backend/requirements.txt
opencv-python
gunicorn
imutils
pydub
moviepy
gensim
IMDbPY
scikit-learn
dlib
face-recognition
SpeechRecognition
flask
mysql-connector-python